Починил скрипты добавления тестовых данных в БД (у всех 4-ех тестовых пользователей пароль "12345678")

+++ db_migrations/changesets/rollback/M-006.rollback.sql
@@ -1,3 +1,4 @@
-DELETE FROM matcher.organizations;
-DELETE FROM matcher.users;
-DELETE FROM matcher.projects;
+DELETE FROM matcher.user_organizations WHERE org_id IN (1, 2, 3, 4);
+DELETE FROM matcher.organizations WHERE org_id IN (1, 2, 3, 4);
+DELETE FROM matcher.users WHERE user_id IN (1, 2, 3, 4);
+DELETE FROM matcher.projects WHERE project_id IN (1, 2, 3);
